Enable permutation of vulnerability functions within groupings
TO allow users to explore the uncertainty associated with the random
distribution of exposure attributes in the exposure database, this version
enables one to randomly permutate the assignment of attributes (and hence
the vulnerability) of assets within a classification group (e.g. suburb,
LGA or meshblock).

- uses `pandas.DataFrame` to move exposure attributes around. The
  `pandas.DataFrame` was chosen because of the ease with which one can
  perform the permutation of attributes, and the ease of aggregating
  resulting loss outcomes.
- the new `wind_v5` template enables the permutation of exposure.
- new jobs and misc functions to support permutation and aggregation.

TODO: Update unit tests