More working tests with pandas DataFrames
Slight modification to the context class to handle different types
for the `exposure_att` attribute (either dict or pandas.DataFrame).

Two tests are skipped at present. The first is
test_load_raster_clipping, as it's failing the comparison of NaN values
(NaN != NaN). The second test relates to loading of multiple raster
files at the same time. Currently not able to insert a 2-d array of
values into a single column of the DataFrame.